+++ lib/src/database/upgrades/force/force_push.dart
@@ -7,6 +7,7 @@ final forcePush = Upgrade(
   isExhaustible: true,
   points: 10,
   text: r'''
-    Action: Choose an enemy trooper unit at range 1. Perform a speed-1 move with that unit, even if it is engaged.
+    Action: Choose an enemy trooper unit at range 1.
+    Perform a speed-1 move with that unit, even if it is engaged.
   ''',
 );