+++ lib/src/database/upgrades/force/jedi_mind_trick.dart
@@ -8,6 +8,7 @@ final jediMindTrick = Upgrade(
   restrictedToFaction: Faction.lightSide,
   points: 5,
   text: r'''
-    Action: Choose an enemy trooper at range 1-2. It gains 2 suppression tokens.
+    Action: Choose an enemy trooper at range 1-2.
+    It gains 2 suppression tokens.
   ''',
 );