+++ lib/src/database/upgrades/gear/environmental_gear.dart
@@ -8,4 +8,7 @@ final environmentalGear = Upgrade(
   text: r'''
     You gain Unhindered.
   ''',
+  keywords: {
+    Keyword.unhindered: '',
+  },
 );