+++ lib/src/database/upgrades/gear/recon_intel.dart
@@ -8,4 +8,7 @@ final reconIntel = Upgrade(
   text: r'''
     You gain Scout 1.
   ''',
+  keywords: {
+    Keyword.scoutX: '1',
+  },
 );