+++ lib/src/database/upgrades/gear/targeting_scopes.dart
@@ -8,4 +8,7 @@ final targetingScopes = Upgrade(
   text: r'''
     You gain Precise 1.
   ''',
+  keywords: {
+    Keyword.preciseX: '1',
+  },
 );